Final version

+++ AbstractFactory.h
@@ -5,8 +5,8 @@
 #include <string>
 
 class AbstractFactory {
-public: 
-    virtual BanPhim *CreateBanPhim(loai Loai, bienthe BienThe, string Ten, string Mau, double Gia, int Sl) const = 0;
-    virtual ManHinh *CreateManHinh(loai Loai, bienthe BienThe, string Ten, string Mau, double Gia, int Sl) const = 0;
-    virtual Chuot *CreateChuot(loai Loai, bienthe BienThe, string Ten, string Mau, double Gia, int Sl) const = 0;
+public:
+    virtual BanPhim* CreateBanPhim(loai Loai, bienthe BienThe, string Ten, string Mau, long Gia, int Sl) const = 0;
+    virtual ManHinh* CreateManHinh(loai Loai, bienthe BienThe, string Ten, string Mau, long Gia, int Sl) const = 0;
+    virtual Chuot* CreateChuot(loai Loai, bienthe BienThe, string Ten, string Mau, long Gia, int Sl) const = 0;
 };
