+++ ToiGianFactory.h
@@ -5,13 +5,13 @@
 
 class ToiGianFactory : public AbstractFactory {
 public:
-    BanPhim *CreateBanPhim(loai Loai, bienthe BienThe, string Ten, string Mau, double Gia, int Sl) const {
+    BanPhim* CreateBanPhim(loai Loai, bienthe BienThe, string Ten, string Mau, long Gia, int Sl) const {
         return new BanPhimToiGian(Loai, BienThe, Ten, Mau, Gia, Sl);
     }
-    ManHinh *CreateManHinh(loai Loai, bienthe BienThe, string Ten, string Mau, double Gia, int Sl) const {
+    ManHinh* CreateManHinh(loai Loai, bienthe BienThe, string Ten, string Mau, long Gia, int Sl) const {
         return new ManHinhToiGian(Loai, BienThe, Ten, Mau, Gia, Sl);
     }
-    Chuot *CreateChuot(loai Loai, bienthe BienThe, string Ten, string Mau, double Gia, int Sl) const {
+    Chuot* CreateChuot(loai Loai, bienthe BienThe, string Ten, string Mau, long Gia, int Sl) const {
         return new ChuotToiGian(Loai, BienThe, Ten, Mau, Gia, Sl);
     }
 };
